+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Find changes for the upcoming release in the project's [changelog.d](https://git
 
 - Prefer environment variables to the mobu YAML configuration file when determining configuration overrides. Otherwise, when mobu is installed by Phalanx using Helm v4, the Sentry environment is overridden to be an empty string.
 
+### Other changes
+
+- mobu now uses [nox](https://nox.thea.codes/en/stable/index.html) as the build system for development and testing instead of tox.
+
 <a id='changelog-19.1.2'></a>
 
 ## 19.1.2 (2026-08-26)